로딩 플레이어 싱크 오류 해결
로딩 플레이어 간 싱크에서 float형을 int형으로 cast하여 발생하는 오류를 해결하였습니다.

+++ Assets/RF/Game/LoadingPlayer.cs
@@ -30,7 +30,7 @@ public class LoadingPlayer : MonoBehaviour, IPunObservable, IPunInstantiateMagic
         }
         else
         {
-            loadingProgress = (int)stream.ReceiveNext();
+            loadingProgress = (float)stream.ReceiveNext();
             
             Debug.Log("Receive Progress : " + loadingProgress);
         }
